community-service/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
isodate==0.6.0
pydantic[email]
python-decouple==3.7
httpx==0.23.3
pandas==1.5.3
python-multipart
bcrypt
redis
Pillow